refactor: update imports to use absolute paths for mpcPeers
- Replaced relative imports of mpcPeers with absolute paths across multiple files for improved clarity and consistency.
- Introduced a new mpcPeers state management file to centralize peer management functionality.
- Updated components to utilize the new structure, ensuring seamless integration with the updated import paths.
