+++ clients/desktop/src/mpc/peers/PeersManagerTitle.tsx
@@ -1,7 +1,6 @@
+import { useMpcPeers } from '@core/ui/mpc/state/mpcPeers'
 import { Text } from '@lib/ui/text'
 import { useTranslation } from 'react-i18next'
-
-import { useMpcPeers } from './state/mpcPeers'
 
 type PeersManagerTitleProps = {
   target: number
