+++ clients/desktop/src/mpc/signers/state/mpcSigners.ts
@@ -1,7 +1,6 @@
 import { useMpcLocalPartyId } from '@core/ui/mpc/state/mpcLocalPartyId'
+import { useMpcPeers } from '@core/ui/mpc/state/mpcPeers'
 import { useMemo } from 'react'
-
-import { useMpcPeers } from '../../peers/state/mpcPeers'
 
 export const useMpcSigners = () => {
   const peers = useMpcPeers()
