+++ clients/desktop/src/vault/keygen/shared/peerDiscovery/KeygenPeerDiscoveryStep.tsx
@@ -3,6 +3,7 @@ import { recommendedPeers, requiredPeers } from '@core/mpc/peers/config'
 import { parseLocalPartyId } from '@core/mpc/signers/localPartyId'
 import { useKeygenVault } from '@core/ui/mpc/keygen/state/keygenVault'
 import { useMpcLocalPartyId } from '@core/ui/mpc/state/mpcLocalPartyId'
+import { useMpcPeers } from '@core/ui/mpc/state/mpcPeers'
 import { Match } from '@lib/ui/base/Match'
 import { InfoIcon } from '@lib/ui/icons/InfoIcon'
 import { OnBackProp, OnForwardProp } from '@lib/ui/props'
@@ -25,7 +26,6 @@ import { PeersContainer } from '../../../../mpc/peers/PeersContainer'
 import { PeersManagerFrame } from '../../../../mpc/peers/PeersManagerFrame'
 import { PeersManagerTitle } from '../../../../mpc/peers/PeersManagerTitle'
 import { PeersPageContentFrame } from '../../../../mpc/peers/PeersPageContentFrame'
-import { useMpcPeers } from '../../../../mpc/peers/state/mpcPeers'
 import { MpcLocalServerIndicator } from '../../../../mpc/serverType/MpcLocalServerIndicator'
 import { useMpcServerType } from '../../../../mpc/serverType/state/mpcServerType'
 import { FitPageContent } from '../../../../ui/page/PageContent'
