+++ clients/desktop/src/vault/keysign/shared/mutations/useKeysignMutation.ts
@@ -11,6 +11,7 @@ import { KeysignMessagePayload } from '@core/mpc/keysign/keysignPayload/KeysignM
 import { MPCKeysign } from '@core/mpc/mpcKeysign'
 import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider'
 import { useIsInitiatingDevice } from '@core/ui/mpc/state/isInitiatingDevice'
+import { useMpcPeers } from '@core/ui/mpc/state/mpcPeers'
 import { useMpcSessionId } from '@core/ui/mpc/state/mpcSession'
 import { getLastItem } from '@lib/utils/array/getLastItem'
 import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
@@ -21,7 +22,6 @@ import { keccak256 } from 'js-sha3'
 
 import { tss } from '../../../../../wailsjs/go/models'
 import { Keysign } from '../../../../../wailsjs/go/tss/TssService'
-import { useMpcPeers } from '../../../../mpc/peers/state/mpcPeers'
 import { useMpcServerUrl } from '../../../../mpc/serverType/state/mpcServerUrl'
 import { getVaultPublicKey } from '../../../publicKey/getVaultPublicKey'
 import { useCurrentHexEncryptionKey } from '../../../setup/state/currentHexEncryptionKey'
